+++ fftAlgorithm/fft.cpp
@@ -19,7 +19,7 @@
 //     Output - transform result
 //     N      - length of both input data and result
 
-bool CFFT::fftAlgorithm(complex *const Input, double *const Output, const unsigned int N)
+bool CFFT::fftAlgorithm(complex *const Input, float *const Output, const unsigned int N)
 {
 	//   Check input parameters
     if (!Input || !Output || N < 1 || N & (N - 1))
@@ -38,9 +38,9 @@ bool CFFT::fftAlgorithm(complex *const Input, double *const Output, const unsign
 }
 
 
- double CFFT::Hamming(unsigned int i, unsigned int N)
+ float CFFT::Hamming(unsigned int i, unsigned int N)
 {
-    double a = 0;
+    float a = 0;
     a =  0.54-0.46*cos(2*M_PI*i/(N-1));
     return a;
 
@@ -55,7 +55,7 @@ bool CFFT::fftAlgorithm(complex *const Input, double *const Output, const unsign
 
 
 //
-void CFFT::countAmplitude(complex *const Input, double *const Output, const unsigned int N)
+void CFFT::countAmplitude(complex *const Input, float *const Output, const unsigned int N)
 {
     unsigned int i = 0;
     for (i = 0; i < N; i++) {
@@ -89,16 +89,16 @@ void CFFT::Rearrange(const complex *const Input, complex *const Output, const un
 //   FFT implementation
 void CFFT::Perform(complex *const Data, const unsigned int N, const bool Inverse /* = false */)
 {
-	const double pi = Inverse ? 3.14159265358979323846 : -3.14159265358979323846;
+    const float pi = Inverse ? 3.14159265358979323846 : -3.14159265358979323846;
 	//   Iteration through dyads, quadruples, octads and so on...
 	for (unsigned int Step = 1; Step < N; Step <<= 1)
 	{
 		//   Jump to the next entry of the same transform factor
 		const unsigned int Jump = Step << 1;
 		//   Angle increment
-		const double delta = pi / double(Step);
+        const float delta = pi / float(Step);
 		//   Auxiliary sin(delta / 2)
-		const double Sine = sin(delta * .5);
+        const float Sine = sin(delta * .5);
 		//   Multiplier for trigonometric recurrence
 		const complex Multiplier(-2. * Sine * Sine, sin(delta));
 		//   Start value for transform factor, fi = 0
